Update sur le tab de Image Inclusion

+++ Dev/LOG792/ImageExtract/ImageExtract/ST/BaseForm.cs
@@ -120,11 +120,11 @@ namespace ImageExtract.ST
 
             // Tab Control 1, Delete button and Save button are only available if a config is selected
             this.tabControl1.Enabled = this.btnDeleteConfig.Enabled = this.btnSaveConfig.Enabled =
-                ((cb.SelectedItem as MyComboBoxOrListBoxItem).Value != NO_CONFIG_SELECTED);
+                ((int)(cb.SelectedItem as MyComboBoxOrListBoxItem).Value != NO_CONFIG_SELECTED);
 
             // If the chosen option isn't "no config selected", remove the "no config selected" choice.
             // We only need that option when entirely reloading the top section, for example at startup
-            if ((cb.SelectedItem as MyComboBoxOrListBoxItem).Value != NO_CONFIG_SELECTED)
+            if ((int)(cb.SelectedItem as MyComboBoxOrListBoxItem).Value != NO_CONFIG_SELECTED)
             {
                 cb.Items.Remove(noConfigSelected);
             }
@@ -132,7 +132,7 @@ namespace ImageExtract.ST
             for (int i = 0; i < this.listOfImageExtractConfigs.Count && !configFoundInList; i++ )
             {
                 if ((listOfImageExtractConfigs[i] as ImageExtractConfig).ImgExtr_Config_Id == 
-                    (this.comboChooseImageExtract.SelectedItem as MyComboBoxOrListBoxItem).Value)
+                    (int)(this.comboChooseImageExtract.SelectedItem as MyComboBoxOrListBoxItem).Value)
                 {
                     currentlyDisplayedImageExtractConfig = listOfImageExtractConfigs[i] as ImageExtractConfig;
                 }
